made project more modular
