+++ prompts/index.js
@@ -1,0 +1,6 @@
+const addDepartment = require('./add_department');
+const addRole = require('./add_role');
+const addEmployee = require('./add_employee');
+const updateEmployee = require('./update_employee');
+
+module.exports = { addDepartment, addRole, addEmployee, updateEmployee };
