+++ queries/get_employee_by_role.js
@@ -1,0 +1,1 @@
+const db = require('../db/db');
