+++ queries/view_departments.js
@@ -1,0 +1,7 @@
+const db = require('../db/db');
+
+db.query('SELECT * FROM departments;', (err, result) => {
+    if (err) throw err;
+
+    console.log(result);
+});
