+++ queries/view_roles.js
@@ -1,0 +1,7 @@
+const db = require('../db/db');
+
+db.query('SELECT * FROM roles;', (err, result) => {
+    if (err) throw err;
+
+    console.log(result);
+});
